[TASK] Support language selection in TYPO3 14 (in page module)
For any reason there is no language variable available if an editors opens a page the first time. So "0" is used here. If the editor switches languages in backend selection, the correct language will be used. Anyway, if the page was selected with "1" and gets reloaded, we use "0" for the moment. Maybe this issue will be fixed in upcoming TYPO3 versions.