[BUGFIX] Use correct language for page module
It turned out that the language is not always given as GET param but stored in be_users.uc field.
I also tried DI via Constructor in the eventlistener but the backenduserauthentication object is then always empty and not related to the currently logged in user. So we have to use GLOBALS again here